+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 -   Fixed Grid not showing last modified date after reopening GUI.
+-   Fixed not being able to extract fluids from External Storage in the Grid on Fabric.
 
 ## [2.0.0-beta.7] - 2025-08-05